+++ deepSculpt/components/planes.py
@@ -58,4 +58,4 @@ def add_plane(
         print_information()
         print("###############################################################")
 
-    return void, color_void
+    return void.astype("int8"), color_void
